Add a left hand navigation bar to navigate between different pages.
